feat: indexing past blocks

+++ internal/models/model.go
@@ -9,9 +9,9 @@ type BaseModel struct {
 
 type Block struct {
 	BaseModel
-	Number       int           `json:"block_num" gorm:"primaryKey;uniqueIndex:,sort:desc"`
+	Number       uint64        `json:"block_num" gorm:"primaryKey;uniqueIndex:,sort:desc"`
 	Hash         string        `json:"block_hash"`
-	Time         int           `json:"block_time"`
+	Time         uint64        `json:"block_time"`
 	ParentHash   string        `json:"parent_hash"`
 	Transactions []Transaction `json:"transactions" gorm:"foreignKey:BlockNumber;references:Number"`
 }
@@ -21,9 +21,9 @@ type Transaction struct {
 	Hash        string `json:"tx_hash" gorm:"primaryKey;type:char(66);uniqueIndex"`
 	From        string `json:"from"`
 	To          string `json:"to"`
-	Nonce       int    `json:"nonce"`
+	Nonce       uint64 `json:"nonce"`
 	Data        string `json:"data"`
 	Value       string `json:"value"`
 	Logs        string `json:"logs"`
-	BlockNumber int    `json:"-"`
+	BlockNumber uint64 `json:"-"`
 }
